delete/userInfo

+++ src/main/java/LD_Caffe/ld_caffe/service/AdminService.java
@@ -31,11 +31,23 @@ public class AdminService {
     }
 
 
-    public Integer deleteUser(String userId){
+    public Integer deleteUser(String userName){
+        System.out.println(userName);
+        // userName로 userId찾아서 그값으로 데이터 삭제
+        Optional<UserEntity> userInfo = userRepository.findByuserName(userName);
+//        List<UserEntity> userInfo = userRepository.findByuserName(userName);
+        if (userInfo.isPresent()){
+            System.out.println("데이터베이스에서 유저명을 찾았습니다.");
+            String userId = userInfo.get().getUserId();
+            userRepository.deleteById(userId);
+            return 1;
+        }else {
+            System.out.println("데이터베이스에서 해당 유저명이 없습니다.");
+            return 0;
+        }
 
-        userRepository.deleteById(userId);
 
-        return 1;
+
     }
 
 }
